+++ app/models/concerns/syncable.rb
@@ -18,14 +18,6 @@ module Syncable
     syncs.create!(start_date: start_date).perform
   end
 
-  def sync_data(sync, start_date: nil)
-    raise NotImplementedError, "Subclasses must implement the `sync_data` method"
-  end
-
-  def post_sync(sync)
-    # no-op, syncable can optionally provide implementation
-  end
-
   def sync_error
     latest_sync&.error
   end
