+++ app/models/family.rb
@@ -35,6 +35,21 @@ class Family < ApplicationRecord
   validates :locale, inclusion: { in: I18n.available_locales.map(&:to_s) }
   validates :date_format, inclusion: { in: DATE_FORMATS.map(&:last) }
 
+  def remove_syncing_notice!
+    broadcast_remove target: "syncing-notice"
+  end
+
+  # If family has any syncs pending/syncing within the last 10 minutes, we show a persistent "syncing" notice.
+  # Ignore syncs older than 10 minutes as they are considered "stale"
+  def syncing?
+    Sync.where(
+      "(syncable_type = 'Family' AND syncable_id = ?) OR
+       (syncable_type = 'Account' AND syncable_id IN (SELECT id FROM accounts WHERE family_id = ? AND plaid_account_id IS NULL)) OR
+       (syncable_type = 'PlaidItem' AND syncable_id IN (SELECT id FROM plaid_items WHERE family_id = ?))",
+      id, id, id
+    ).where(status: [ "pending", "syncing" ], created_at: 10.minutes.ago..).exists?
+  end
+
   def assigned_merchants
     merchant_ids = transactions.where.not(merchant_id: nil).pluck(:merchant_id).uniq
     Merchant.where(id: merchant_ids)
