+++ app/views/accounts/_accountable_group.html.erb
@@ -15,11 +15,11 @@
 
   <div class="space-y-1">
     <% account_group.accounts.each do |account| %>
-      <%= link_to account_path(account), 
+      <%= link_to account_path(account),
                   class: class_names(
                     "block flex items-center gap-2 px-3 py-2 rounded-lg",
                     page_active?(account_path(account)) ? "bg-container" : "hover:bg-surface-hover"
-                  ), 
+                  ),
                   data: { sidebar_tabs_target: "account", action: "click->sidebar-tabs#select" },
                   title: account.name do %>
         <%= render "accounts/logo", account: account, size: "sm", color: account_group.color %>
