+++ app/views/category/dropdowns/show.html.erb
@@ -2,13 +2,13 @@
   <div class="flex flex-col relative" data-controller="list-filter">
     <div class="grow p-1.5">
       <div class="relative flex items-center bg-container border border-secondary rounded-lg">
-        <input 
-          placeholder="<%= t(".search_placeholder") %>" 
-          autocomplete="nope" 
-          type="search" 
-          class="bg-container placeholder:text-sm placeholder:text-secondary font-normal h-10 relative pl-10 w-full border-none rounded-lg focus:outline-hidden focus:ring-0" 
-          data-list-filter-target="input" 
-          data-action="list-filter#filter" />
+        <input
+          placeholder="<%= t(".search_placeholder") %>"
+          autocomplete="nope"
+          type="search"
+          class="bg-container placeholder:text-sm placeholder:text-secondary font-normal h-10 relative pl-10 w-full border-none rounded-lg focus:outline-hidden focus:ring-0"
+          data-list-filter-target="input"
+          data-action="list-filter#filter">
         <%= icon("search", class: "absolute inset-0 ml-2 transform top-1/2 -translate-y-1/2") %>
       </div>
     </div>
